clear controllers
